fix cuda dev_ctx allocator cmake deps, test=develop
