+++ paddle/fluid/memory/malloc.cc
@@ -17,10 +17,6 @@ limitations under the License. */
 #include <vector>
 #include "paddle/fluid/memory/allocation/allocator_facade.h"
 #include "paddle/fluid/memory/allocation/allocator_strategy.h"
-#ifdef PADDLE_WITH_CUDA
-#include "paddle/fluid/memory/allocation/cuda_device_context_allocator_pool.h"
-#endif
-#include "paddle/fluid/platform/device_context.h"
 #include "paddle/fluid/platform/place.h"
 
 namespace paddle {
@@ -33,27 +29,6 @@ std::shared_ptr<Allocation> AllocShared(const platform::Place &place,
 
 AllocationPtr Alloc(const platform::Place &place, size_t size) {
   return allocation::AllocatorFacade::Instance().Alloc(place, size);
-}
-
-AllocationPtr Alloc(const platform::DeviceContext &dev_ctx, size_t size) {
-  auto place = dev_ctx.GetPlace();
-#ifdef PADDLE_WITH_CUDA
-  if (size == 0 || !platform::is_gpu_place(place)) {
-    return Alloc(place, size);
-  }
-  auto *default_dev_ctx = static_cast<platform::CUDADeviceContext *>(
-      platform::DeviceContextPool::Instance().Get(place));
-  auto &desired_dev_ctx =
-      static_cast<const platform::CUDADeviceContext &>(dev_ctx);
-  if (default_dev_ctx->stream() == desired_dev_ctx.stream()) {
-    return Alloc(place, size);
-  } else {
-    return allocation::CUDADeviceContextAllocatorPool::Instance().Alloc(
-        desired_dev_ctx, size);
-  }
-#else
-  return Alloc(place, size);
-#endif
 }
 
 }  // namespace memory
