+++ paddle/fluid/platform/device_context.h
@@ -220,14 +220,7 @@ class CudnnWorkspaceHandle {
     ResetWorkspace();
   }
 
-  inline void ReallocWorkspace(size_t required_workspace_bytes) {
-    if (required_workspace_bytes <= WorkspaceSize()) {
-      return;
-    }
-    // reset allocation first before re-allocate to save memory
-    allocation_.reset();
-    allocation_ = memory::Alloc(device_context_, required_workspace_bytes);
-  }
+  void ReallocWorkspace(size_t required_workspace_bytes);
 
   inline void ResetWorkspace() { allocation_ = nullptr; }
